Refactor dashboard utilities into package
