+++ focus_monitor_dashboard.py
@@ -12,18 +12,21 @@ import pandas as pd
 import plotly.express as px
 import plotly.graph_objects as go
 import psutil
-import requests
 import streamlit as st
 
+from dashboard.data_utils import (
+    load_available_dates, load_log_entries, load_daily_summary, generate_summary_from_logs,
+    load_labels, save_labels, apply_labels_to_logs, group_activities_for_labeling,
+    load_block_feedback, save_block_feedback, load_time_buckets_for_date,
+    update_bucket_summary_in_file, load_categories, save_categories,
+    update_bucket_category_in_file, generate_time_buckets_from_logs,
+)
+from dashboard.llm_utils import (
+    _call_llm_api, generate_summary_and_category, generate_summary_from_raw_with_llm,
+    refine_summary_with_llm, process_activity_data,
+)
+from dashboard.charts import create_pie_chart, create_browser_chart, create_category_chart
 # --- Configuration ---
-LOGS_DIR = Path(__file__).parent / "focus_logs"
-LABELS_FILE = LOGS_DIR / "activity_labels.json"
-FEEDBACK_FILE = LOGS_DIR / "block_feedback.json"  # For personal notes
-CATEGORIES_FILE = LOGS_DIR / "activity_categories.json"  # For user-defined categories
-
-LLM_API_URL = "http://localhost:11434/api/generate"  # Ollama API endpoint
-LLM_MODEL = "llama3.1:8b"  # Specify your desired model
-
 
 # --- Tracker Control Functions ---
 def is_tracker_running():
@@ -354,1319 +357,6 @@ def llm_test_page():
 
 
 # --- Data Management Functions ---
-def load_available_dates():
-    LOGS_DIR.mkdir(parents=True, exist_ok=True)
-    log_dates = set(
-        [
-            f.name.split("_")[-1].replace(".jsonl", "")
-            for f in LOGS_DIR.glob("focus_log_*.jsonl")
-        ]
-    )
-    summary_dates = set(
-        [
-            f.name.split("_")[-1].replace(".json", "")
-            for f in LOGS_DIR.glob("daily_summary_*.json")
-        ]
-    )
-    all_dates = sorted(list(log_dates.union(summary_dates)), reverse=True)
-    return all_dates
-
-
-def load_log_entries(date_str: str) -> pd.DataFrame:
-    file_path = LOGS_DIR / f"focus_log_{date_str}.jsonl"
-    if not file_path.exists():
-        return pd.DataFrame()
-    data = []
-    with open(file_path, "r", encoding="utf-8") as f:
-        for line in f:
-            try:
-                entry = json.loads(line.strip())
-                data.append(entry)
-            except json.JSONDecodeError:
-                # Optionally log this error
-                continue
-    return pd.DataFrame(data)
-
-
-def load_daily_summary(date_str: str) -> Optional[Dict[str, Any]]:
-    file_path = LOGS_DIR / f"daily_summary_{date_str}.json"
-    if file_path.exists():
-        try:
-            with open(file_path, "r", encoding="utf-8") as f:
-                return json.load(f)
-        except json.JSONDecodeError:
-            st.error(
-                f"Error decoding summary file for {date_str}. Generating from logs if possible."
-            )
-
-    log_path = LOGS_DIR / f"focus_log_{date_str}.jsonl"
-    if log_path.exists():
-        st.info(f"No pre-generated summary for {date_str}. Generating from raw logs...")
-        return generate_summary_from_logs(date_str)
-    else:
-        st.warning(f"No log data available to generate summary for {date_str}")
-        return None
-
-
-def generate_summary_from_logs(date_str: str) -> Optional[Dict[str, Any]]:
-    logs_df = load_log_entries(date_str)
-    if logs_df.empty:
-        return None
-
-    logs_df = apply_labels_to_logs(logs_df)  # Apply custom labels first
-
-    total_duration = logs_df.get("duration", pd.Series(dtype="float")).sum()
-    summary = {"date": date_str, "totalTime": total_duration, "appBreakdown": []}
-    if total_duration == 0:
-        return summary
-
-    # Ensure 'app_name' exists before grouping; default if not
-    if "app_name" not in logs_df.columns:
-        logs_df["app_name"] = "Unknown Application"
-    if "exe" not in logs_df.columns:
-        logs_df["exe"] = "unknown.exe"
-
-    app_groups = (
-        logs_df.groupby("app_name")
-        .agg(
-            exe_path=("exe", "first"),
-            time_spent=("duration", "sum"),
-            window_titles=(
-                ("title", lambda x: list(set(str(t) for t in x if pd.notna(t)))[:50])
-                if "title" in logs_df.columns
-                else ([],)
-            ),
-        )
-        .reset_index()
-    )
-
-    for _, row in app_groups.iterrows():
-        percentage = (
-            (row["time_spent"] / total_duration * 100) if total_duration > 0 else 0
-        )
-        summary["appBreakdown"].append(
-            {
-                "appName": row["app_name"],
-                "exePath": row.get("exe_path", "unknown.exe"),
-                "timeSpent": int(row["time_spent"]),
-                "percentage": round(percentage, 2),
-                "windowTitles": (
-                    row.get("window_titles", [])
-                    if isinstance(row.get("window_titles"), list)
-                    else [str(row.get("window_titles"))]
-                ),
-            }
-        )
-
-    summary["appBreakdown"].sort(key=lambda x: x["timeSpent"], reverse=True)
-    return summary
-
-
-# --- Label Management Functions ---
-def load_labels() -> Dict[str, Any]:
-    if not LABELS_FILE.exists():
-        return {}
-    try:
-        with open(LABELS_FILE, "r", encoding="utf-8") as f:
-            return json.load(f)
-    except Exception as e:
-        st.error(f"Error loading labels: {e}")
-        return {}
-
-
-def save_labels(labels: Dict[str, Any]) -> bool:
-    try:
-        LOGS_DIR.mkdir(parents=True, exist_ok=True)
-        with open(LABELS_FILE, "w", encoding="utf-8") as f:
-            json.dump(labels, f, indent=2)
-        return True
-    except Exception as e:
-        st.error(f"Error saving labels: {e}")
-        return False
-
-
-def apply_labels_to_logs(logs_df: pd.DataFrame) -> pd.DataFrame:
-    if logs_df.empty:
-        return logs_df
-
-    labels = load_labels()
-    if not labels:  # No custom labels defined
-        df_copy = logs_df.copy()
-        df_copy["original_app_name"] = df_copy.get("app_name", "Unknown Application")
-        return df_copy
-
-    df = logs_df.copy()
-
-    # Ensure essential columns exist and handle NaN by converting to string
-    for col, default_val in [
-        ("app_name", "Unknown Application"),
-        ("exe", "unknown.exe"),
-        ("title", ""),
-    ]:
-        if col not in df.columns:
-            df[col] = default_val
-        df[col] = df[col].fillna(default_val).astype(str)
-
-    df["original_app_name"] = df["app_name"]  # Preserve original before modification
-    df["labeled_app_name"] = df[
-        "app_name"
-    ]  # Start with current app_name, then override
-
-    # Apply exact titles: key is "exact::FULL_EXE_PATH::TITLE_STRING"
-    for key, label_val in labels.get("exact_titles", {}).items():
-        try:
-            _, exe_from_key, title_from_key = key.split("::", 2)
-            mask = (df["exe"].str.lower() == exe_from_key.lower()) & (
-                df["title"] == title_from_key
-            )  # Case-sensitive title match
-            df.loc[mask, "labeled_app_name"] = label_val
-        except ValueError:
-            continue  # Skip malformed keys
-        except Exception as e:
-            print(f"Error applying exact title label for '{key}': {e}")
-
-    # Apply exact exe (basename): key is "BASENAME.exe"
-    for exe_basename_key, label_val in labels.get("exact_exe", {}).items():
-        # Apply only if not already labeled by a more specific rule (exact_title)
-        mask = (
-            df["exe"].apply(lambda x: os.path.basename(x).lower())
-            == exe_basename_key.lower()
-        ) & (df["labeled_app_name"] == df["original_app_name"])
-        df.loc[mask, "labeled_app_name"] = label_val
-
-    # Apply patterns (basename): key is "pattern::BASENAME.exe::TITLE_PATTERN"
-    for key, label_val in labels.get("patterns", {}).items():
-        try:
-            _, exe_basename_key, pattern_from_key = key.split("::", 2)
-            # Apply only if not already labeled by a more specific rule
-            mask = (
-                (
-                    df["exe"]
-                    .apply(lambda x: os.path.basename(x).lower())
-                    .str.contains(exe_basename_key.lower(), regex=False)
-                )
-                & (
-                    df["title"]
-                    .str.lower()
-                    .str.contains(pattern_from_key.lower(), regex=False)
-                )
-                & (df["labeled_app_name"] == df["original_app_name"])
-            )
-            df.loc[mask, "labeled_app_name"] = label_val
-        except ValueError:
-            continue
-        except Exception as e:
-            print(f"Error applying pattern label for '{key}': {e}")
-
-    df["app_name"] = df["labeled_app_name"]  # Final app_name is the labeled one
-    df.drop(columns=["labeled_app_name"], inplace=True, errors="ignore")
-    return df
-
-
-def group_activities_for_labeling(logs_df: pd.DataFrame) -> pd.DataFrame:
-    if logs_df.empty or "title" not in logs_df.columns:
-        return pd.DataFrame()
-
-    # Ensure necessary columns exist and handle NaNs
-    for col, default_val in [
-        ("app_name", "Unknown Application"),
-        ("exe", "unknown.exe"),
-        ("duration", 0),
-    ]:
-        if col not in logs_df.columns:
-            logs_df[col] = default_val
-        logs_df[col] = logs_df[col].fillna(default_val)
-
-    logs_df["title"] = logs_df["title"].astype(
-        str
-    )  # Ensure title is string for grouping
-
-    # Group by title, then find mode for app_name and exe, sum duration
-    title_groups_data = []
-    for title_val, group in logs_df.groupby("title"):
-        app_name_mode = group["app_name"].mode()
-        exe_mode = group["exe"].mode()
-        title_groups_data.append(
-            {
-                "title": title_val,
-                "app_name": (
-                    app_name_mode[0]
-                    if not app_name_mode.empty
-                    else "Unknown Application"
-                ),
-                "exe": (
-                    exe_mode[0] if not exe_mode.empty else "unknown.exe"
-                ),  # Store full path
-                "count": len(group),
-                "duration": group["duration"].sum(),
-            }
-        )
-
-    grouped_df = pd.DataFrame(title_groups_data)
-    if not grouped_df.empty:
-        total_log_duration = logs_df["duration"].sum()
-        grouped_df["percentage"] = (
-            (grouped_df["duration"] / total_log_duration * 100).round(1)
-            if total_log_duration > 0
-            else 0.0
-        )
-        return grouped_df.sort_values("duration", ascending=False)
-    return pd.DataFrame()
-
-
-# --- Feedback Management Functions (for block_feedback.json - Personal Notes) ---
-def load_block_feedback() -> Dict[str, str]:
-    if not FEEDBACK_FILE.exists():
-        return {}
-    try:
-        with open(FEEDBACK_FILE, "r", encoding="utf-8") as f:
-            return json.load(f)
-    except Exception as e:
-        st.error(f"Error loading personal notes (feedback file): {e}")
-        return {}
-
-
-def save_block_feedback(feedback_data: Dict[str, str]) -> bool:
-    try:
-        LOGS_DIR.mkdir(parents=True, exist_ok=True)
-        with open(FEEDBACK_FILE, "w", encoding="utf-8") as f:
-            json.dump(feedback_data, f, indent=2)
-        return True
-    except Exception as e:
-        st.error(f"Error saving personal notes (feedback file): {e}")
-        return False
-
-
-# --- Time Bucket Management Functions (for time_buckets_*.json files - Official Summaries) ---
-def load_time_buckets_for_date(date_str: str) -> List[Dict[str, Any]]:
-    buckets_with_tags = []
-    for path in LOGS_DIR.glob("time_buckets_*.json"):
-        try:
-            session_tag = path.name.replace("time_buckets_", "").replace(".json", "")
-            with open(path, "r", encoding="utf-8") as f:
-                data = json.load(f)
-            for b in data:
-                if b.get("start", "")[:10] == date_str:
-                    b_copy = b.copy()
-                    b_copy["session_tag"] = (
-                        session_tag  # Crucial for identifying source file
-                    )
-                    buckets_with_tags.append(b_copy)
-        except Exception as e:
-            st.error(f"Error reading bucket file {path.name}: {e}")
-    buckets_with_tags.sort(key=lambda x: x.get("start", ""))
-    return buckets_with_tags
-
-
-def update_bucket_summary_in_file(
-    session_tag: str, bucket_start_iso: str, new_summary: str
-) -> bool:
-    bucket_file_path = LOGS_DIR / f"time_buckets_{session_tag}.json"
-    if not bucket_file_path.exists():
-        st.error(f"Official summary file not found: {bucket_file_path.name}")
-        return False
-    try:
-        with open(bucket_file_path, "r", encoding="utf-8") as f:
-            all_buckets_data = json.load(f)
-        found_bucket = False
-        for bucket_data in all_buckets_data:
-            if bucket_data.get("start") == bucket_start_iso:
-                bucket_data["summary"] = new_summary
-                found_bucket = True
-                break
-        if not found_bucket:
-            st.error(f"Bucket {bucket_start_iso} not found in {bucket_file_path.name}")
-            return False
-        with open(bucket_file_path, "w", encoding="utf-8") as f:
-            json.dump(all_buckets_data, f, indent=2)
-        return True
-    except Exception as e:
-        st.error(f"Error updating official summary in {bucket_file_path.name}: {e}")
-        return False
-
-
-# --- Category Management Functions ---
-def load_categories():
-    """Load user-defined activity categories from JSON file."""
-    categories_file = Path(__file__).parent / "focus_logs" / "activity_categories.json"
-    if not categories_file.exists():
-        return []
-    try:
-        with open(categories_file, "r", encoding="utf-8") as f:
-            data = json.load(f)
-            return data.get("categories", [])
-    except Exception as e:
-        print(f"Error loading categories: {e}")
-        return []
-
-
-def save_categories(categories: List[Dict[str, str]]) -> bool:
-    """Save user-defined activity categories to JSON file."""
-    try:
-        LOGS_DIR.mkdir(parents=True, exist_ok=True)
-        with open(CATEGORIES_FILE, "w", encoding="utf-8") as f:
-            json.dump({"categories": categories}, f, indent=2)
-        return True
-    except Exception as e:
-        st.error(f"Error saving categories: {e}")
-        return False
-
-
-def update_bucket_category_in_file(
-    session_tag: str, bucket_start_iso: str, new_category_id: str
-) -> bool:
-    """Update the category_id field for a specific time bucket in its file."""
-    bucket_file_path = LOGS_DIR / f"time_buckets_{session_tag}.json"
-    if not bucket_file_path.exists():
-        st.error(f"Official summary file not found: {bucket_file_path.name}")
-        return False
-    try:
-        with open(bucket_file_path, "r", encoding="utf-8") as f:
-            all_buckets_data = json.load(f)
-        found_bucket = False
-        for bucket_data in all_buckets_data:
-            if bucket_data.get("start") == bucket_start_iso:
-                bucket_data["category_id"] = new_category_id
-                found_bucket = True
-                break
-        if not found_bucket:
-            st.error(f"Bucket {bucket_start_iso} not found in {bucket_file_path.name}")
-            return False
-        with open(bucket_file_path, "w", encoding="utf-8") as f:
-            json.dump(all_buckets_data, f, indent=2)
-        return True
-    except Exception as e:
-        st.error(f"Error updating bucket category in {bucket_file_path.name}: {e}")
-        return False
-
-
-# --- LLM Interaction Functions ---
-def _call_llm_api(prompt: str, operation_name: str = "LLM call") -> Optional[str]:
-    """Generic LLM API call helper."""
-    LLM_API_URL = "http://localhost:11434/api/generate"
-    LLM_MODEL = "llama3.1:8b"
-
-    payload = {"model": LLM_MODEL, "prompt": prompt, "stream": False}
-    try:
-        print(f"Sending request to LLM ({LLM_MODEL}) for {operation_name}...")
-        response = requests.post(LLM_API_URL, json=payload, timeout=60)
-        response.raise_for_status()
-        data = response.json()
-        llm_response_text = data.get("response", "").strip()
-        if not llm_response_text and prompt:
-            print(f"LLM returned an empty response for {operation_name}.")
-        else:
-            print(f"LLM {operation_name} successful.")
-        return llm_response_text
-    except requests.exceptions.Timeout:
-        print(f"LLM API request timed out during {operation_name}.")
-    except requests.exceptions.RequestException as e:
-        print(f"LLM API communication error during {operation_name}: {e}")
-    except Exception as e:
-        print(f"An unexpected error occurred during LLM {operation_name}: {e}")
-    return None
-
-
-# --- Add these retroactive processing functions ---
-
-
-def display_retroactive_processor():
-    """Tab for retroactively processing historical focus logs."""
-    st.title("⏮️ Historical Data Processor")
-
-    # Check available dates with raw logs
-    all_dates = load_available_dates()
-
-    # Find dates with logs but missing time buckets or summaries
-    incomplete_dates = []
-
-    for date_str in all_dates:
-        log_path = LOGS_DIR / f"focus_log_{date_str}.jsonl"
-        summary_path = LOGS_DIR / f"daily_summary_{date_str}.json"
-
-        # Check if time buckets exist for this date
-        time_buckets = load_time_buckets_for_date(date_str)
-
-        if log_path.exists() and (not summary_path.exists() or not time_buckets):
-            incomplete_dates.append(
-                {
-                    "date": date_str,
-                    "has_summary": summary_path.exists(),
-                    "has_buckets": bool(time_buckets),
-                    "bucket_count": len(time_buckets) if time_buckets else 0,
-                }
-            )
-
-    if not incomplete_dates:
-        st.success(
-            "All focus logs have been processed! No historical data needs processing."
-        )
-        st.info(
-            "If you've just added categories and want to apply them to existing time buckets, use the 'Recategorize Existing Data' section below."
-        )
-    else:
-        st.warning(
-            f"Found {len(incomplete_dates)} dates with focus logs that need processing."
-        )
-
-        # Display dates needing processing
-        df_incomplete = pd.DataFrame(incomplete_dates)
-        st.dataframe(df_incomplete, use_container_width=True)
-
-        # Process selected date
-        selected_date = st.selectbox(
-            "Select a date to process",
-            options=[d["date"] for d in incomplete_dates],
-            key="retroactive_date_select",
-        )
-
-        selected_info = next(
-            (d for d in incomplete_dates if d["date"] == selected_date), None
-        )
-
-        if selected_info:
-            col1, col2 = st.columns(2)
-
-            # Generate summary if needed
-            if not selected_info["has_summary"]:
-                if col1.button("Generate Daily Summary", key="gen_daily_summary_btn"):
-                    with st.spinner(f"Generating daily summary for {selected_date}..."):
-                        summary = generate_summary_from_logs(selected_date)
-                        if summary:
-                            # Save summary
-                            LOGS_DIR.mkdir(parents=True, exist_ok=True)
-                            summary_path = (
-                                LOGS_DIR / f"daily_summary_{selected_date}.json"
-                            )
-                            with open(summary_path, "w", encoding="utf-8") as f:
-                                json.dump(summary, f, indent=2)
-                            st.success(
-                                f"Daily summary for {selected_date} generated and saved!"
-                            )
-                            time.sleep(1)
-                            st.rerun()
-                        else:
-                            st.error(f"Failed to generate summary for {selected_date}")
-
-            # Generate time buckets if needed
-            if not selected_info["has_buckets"] or selected_info["bucket_count"] == 0:
-                if col2.button("Generate Time Buckets", key="gen_time_buckets_btn"):
-                    with st.spinner(f"Generating time buckets for {selected_date}..."):
-                        success = generate_time_buckets_from_logs(selected_date)
-                        if success:
-                            st.success(
-                                f"Time buckets for {selected_date} generated and saved!"
-                            )
-                            time.sleep(1)
-                            st.rerun()
-                        else:
-                            st.error(
-                                f"Failed to generate time buckets for {selected_date}"
-                            )
-
-    # Section for recategorizing existing time buckets
-    st.markdown("---")
-    st.subheader("Recategorize Existing Data")
-    st.info(
-        "This will apply your current categories to existing time buckets using the LLM."
-    )
-
-    # Get all dates with time buckets
-    dates_with_buckets = []
-    for date_str in all_dates:
-        buckets = load_time_buckets_for_date(date_str)
-        if buckets:
-            # Count buckets with and without categories
-            with_category = sum(1 for b in buckets if b.get("category_id", ""))
-            total_buckets = len(buckets)
-
-            dates_with_buckets.append(
-                {
-                    "date": date_str,
-                    "total_buckets": total_buckets,
-                    "categorized_buckets": with_category,
-                    "needs_categorization": with_category < total_buckets,
-                }
-            )
-
-    if not dates_with_buckets:
-        st.warning("No time bucket data found.")
-    else:
-        # Display dates with their categorization status
-        df_dates = pd.DataFrame(dates_with_buckets)
-        df_dates["Categorization %"] = (
-            df_dates["categorized_buckets"] / df_dates["total_buckets"] * 100
-        ).round(1)
-        st.dataframe(df_dates, use_container_width=True)
-
-        # Select date to recategorize
-        recategorize_date = st.selectbox(
-            "Select a date to recategorize",
-            options=[d["date"] for d in dates_with_buckets],
-            key="recategorize_date_select",
-        )
-
-        # Get buckets for the selected date
-        buckets_to_recategorize = load_time_buckets_for_date(recategorize_date)
-
-        # Count uncategorized buckets
-        uncategorized_count = sum(
-            1 for b in buckets_to_recategorize if not b.get("category_id", "")
-        )
-
-        # Options for recategorization
-        recategorize_all = st.checkbox(
-            "Recategorize all buckets (even those already categorized)", value=False
-        )
-
-        if st.button("Apply Categories to Time Buckets", key="apply_categories_btn"):
-            # Load categories
-            categories = load_categories()
-            if not categories:
-                st.error("No categories defined. Please create categories first.")
-            else:
-                bucket_count = len(buckets_to_recategorize)
-
-                # Create progress bar
-                progress_bar = st.progress(0)
-                status_text = st.empty()
-
-                processed = 0
-                success_count = 0
-
-                # Process each bucket
-                for bucket in buckets_to_recategorize:
-                    # Skip already categorized buckets unless recategorize_all is checked
-                    if bucket.get("category_id", "") and not recategorize_all:
-                        processed += 1
-                        progress_bar.progress(processed / bucket_count)
-                        continue
-
-                    bucket_start_iso = bucket.get("start", "")
-                    session_tag = bucket.get("session_tag", "")
-
-                    status_text.text(
-                        f"Processing bucket {processed+1}/{bucket_count}: {bucket_start_iso}"
-                    )
-
-                    # Get summary and raw data
-                    summary = bucket.get("summary", "")
-                    raw_titles = bucket.get("titles", [])
-                    raw_ocr = bucket.get("ocr_text", [])
-
-                    if summary:
-                        # Use the summary to determine category
-                        # Fix: Unpack 3 values instead of 2
-                        _, category_id, _ = refine_summary_with_llm(
-                            summary, "Please categorize this activity.", ""
-                        )
-                    elif raw_titles or raw_ocr:
-                        # Generate new summary and category from raw data
-                        # Fix: Unpack 3 values instead of 2
-                        summary, category_id, _ = generate_summary_from_raw_with_llm(
-                            raw_titles, raw_ocr
-                        )
-                    else:
-                        # No data to categorize
-                        category_id = ""
-
-                    if category_id:
-                        # Update category in the file
-                        if update_bucket_category_in_file(
-                            session_tag, bucket_start_iso, category_id
-                        ):
-                            success_count += 1
-
-                    processed += 1
-                    progress_bar.progress(processed / bucket_count)
-
-                if success_count > 0:
-                    st.success(
-                        f"Successfully categorized {success_count} buckets for {recategorize_date}!"
-                    )
-                else:
-                    st.warning(
-                        "No buckets were categorized. They may already have categories or lack data."
-                    )
-
-                time.sleep(1)
-                st.rerun()
-
-
-def generate_time_buckets_from_logs(date_str: str) -> bool:
-    """Generate time buckets from focus logs for a specific date."""
-    log_path = LOGS_DIR / f"focus_log_{date_str}.jsonl"
-    if not log_path.exists():
-        st.error(f"Focus log for {date_str} not found.")
-        return False
-
-    # Load raw logs
-    logs_df = load_log_entries(date_str)
-    if logs_df.empty:
-        st.error(f"No log entries found for {date_str}.")
-        return False
-
-    # Ensure timestamp column exists and convert to datetime
-    if "timestamp" not in logs_df.columns:
-        st.error("Log data missing timestamp column.")
-        return False
-
-    logs_df["timestamp"] = pd.to_datetime(logs_df["timestamp"])
-
-    # Sort by timestamp
-    logs_df = logs_df.sort_values("timestamp")
-
-    # Get min and max timestamp for the day
-    min_time = logs_df["timestamp"].min()
-    max_time = logs_df["timestamp"].max()
-
-    # Create 5-minute buckets
-    bucket_size = pd.Timedelta(minutes=5)
-
-    # Generate bucket start times (floor to 5-minute intervals)
-    start_time = min_time.floor("5min")
-    end_time = max_time.ceil("5min")
-
-    bucket_starts = []
-    current = start_time
-    while current <= end_time:
-        bucket_starts.append(current)
-        current += bucket_size
-
-    # Create empty buckets
-    buckets = []
-    session_tag = datetime.now().strftime("%Y%m%dT%H%M%SZ")  # Unique session tag
-
-    for i in range(len(bucket_starts) - 1):
-        bucket_start = bucket_starts[i]
-        bucket_end = bucket_starts[i + 1]
-
-        # Filter logs for this bucket
-        bucket_logs = logs_df[
-            (logs_df["timestamp"] >= bucket_start) & (logs_df["timestamp"] < bucket_end)
-        ]
-
-        if not bucket_logs.empty:
-            # Collect window titles
-            titles = []
-            if "title" in bucket_logs.columns:
-                titles = bucket_logs["title"].dropna().unique().tolist()
-
-            # Create bucket
-            bucket = {
-                "start": bucket_start.isoformat(),
-                "end": bucket_end.isoformat(),
-                "titles": titles,
-                "ocr_text": [],  # Empty as historical logs might not have OCR
-            }
-
-            # Generate summary and category with LLM
-            # Fix: Unpack 3 values instead of 2
-            summary, category_id, _ = generate_summary_from_raw_with_llm(titles, [])
-            if summary:
-                bucket["summary"] = summary
-            if category_id:
-                bucket["category_id"] = category_id
-
-            buckets.append(bucket)
-
-    if not buckets:
-        st.warning(f"No activity buckets created for {date_str}.")
-        return False
-
-    # Save buckets to file
-    bucket_file_path = LOGS_DIR / f"time_buckets_{session_tag}.json"
-    try:
-        with open(bucket_file_path, "w", encoding="utf-8") as f:
-            json.dump(buckets, f, indent=2)
-        return True
-    except Exception as e:
-        st.error(f"Error saving time buckets: {e}")
-        return False
-
-
-def process_activity_data(
-    bucket_titles: List[str],
-    bucket_ocr_texts: List[str],
-    allow_suggestions: bool = True,
-    update_bucket: bool = False,
-    bucket_data: Optional[Dict[str, Any]] = None,
-    bucket_file_path: Optional[str] = None,
-    bucket_index: Optional[int] = None,
-) -> Tuple[Optional[str], Optional[str], Optional[str]]:
-    """
-    Unified function to generate summaries/categories from activity data and optionally update bucket files.
-
-    This function combines the functionality of generate_summary_from_raw_with_llm and categorize_new_time_bucket.
-
-    Args:
-        bucket_titles: List of window titles
-        bucket_ocr_texts: List of OCR text snippets
-        allow_suggestions: Whether to allow the LLM to suggest new categories
-        update_bucket: Whether to update a bucket file with the generated data
-        bucket_data: The bucket data to update (required if update_bucket=True)
-        bucket_file_path: Path to the bucket file (required if update_bucket=True)
-        bucket_index: Index of bucket in the file (required if update_bucket=True)
-
-    Returns:
-        Tuple of (summary_text, category_id, suggested_category)
-    """
-    # Validate and clean inputs
-    titles_to_send = (
-        [str(t) for t in bucket_titles if t and str(t).strip()] if bucket_titles else []
-    )
-    ocr_to_send = (
-        [str(o) for o in bucket_ocr_texts if o and str(o).strip()]
-        if bucket_ocr_texts
-        else []
-    )
-
-    text_parts = list(set(titles_to_send + ocr_to_send))  # Unique raw data points
-    if not text_parts:
-        print("No raw titles or OCR text available to generate a summary")
-        return "", "", ""  # Return empty strings for summary, category, and suggestion
-
-    # Check if update_bucket is True but required params are missing
-    if update_bucket and (
-        bucket_data is None or bucket_file_path is None or bucket_index is None
-    ):
-        print("Warning: Cannot update bucket - missing required parameters")
-        update_bucket = False
-
-    # Get available categories
-    categories = load_categories()
-
-    # Build the prompt for the LLM
-    prompt_text = "Please provide a concise summary of computer activity based on the following window titles and detected text fragments. Focus on the primary tasks or topics.\n\n"
-
-    if titles_to_send:
-        prompt_text += (
-            "Window Titles:\n"
-            + "\n".join([f'- "{t}"' for t in titles_to_send])
-            + "\n\n"
-        )
-    if ocr_to_send:
-        prompt_text += (
-            "Detected Text (OCR Snippets):\n"
-            + "\n".join([f'- "{o}"' for o in ocr_to_send])
-            + "\n\n"
-        )
-
-    # Add category classification if categories exist
-    if categories:
-        prompt_text += "Based on the activity, please categorize it into ONE of the following categories:\n\n"
-        for cat in categories:
-            prompt_text += (
-                f"- {cat.get('name')} ({cat.get('id')}): {cat.get('description')}\n"
-            )
-
-        if allow_suggestions:
-            prompt_text += "\nIf none of these categories fit well, you can suggest a new category instead."
-            prompt_text += "\n\nFirst, provide a concise summary of the activity."
-            prompt_text += "\nThen, on a new line after 'CATEGORY:', provide ONLY the category ID that best matches the activity, or 'none' if no categories fit well."
-            prompt_text += "\nIf you answered 'none', then on a new line after 'SUGGESTION:', provide a suggested new category name and short description in the format 'name | description'."
-        else:
-            prompt_text += "\nFirst, provide a concise summary of the activity. Then, on a new line after 'CATEGORY:', provide ONLY the category ID that best matches the activity."
-    else:
-        prompt_text += "Output only the summary text."
-
-    # Call LLM API
-    llm_response = _call_llm_api(prompt_text, "activity processing")
-
-    if not llm_response:
-        return "", "", ""
-
-    # Parse response to extract summary, category, and suggestion
-    summary_text = llm_response
-    category_id = ""
-    suggested_category = ""
-
-    if categories:
-        parts = llm_response.split("CATEGORY:")
-        if len(parts) > 1:
-            summary_text = parts[0].strip()
-            category_text = parts[1].strip()
-
-            # Check for suggestion
-            suggestion_parts = category_text.split("SUGGESTION:")
-            if len(suggestion_parts) > 1:
-                category_text = suggestion_parts[0].strip()
-                suggested_category = suggestion_parts[1].strip()
-
-            # Extract just the category ID (assuming it's the first word after "CATEGORY:")
-            category_id = category_text.split()[0] if category_text.split() else ""
-
-            # Handle case where LLM says 'none'
-            if category_id.lower() == "none":
-                category_id = ""
-            else:
-                # Verify category ID exists
-                valid_cat_ids = [cat.get("id", "") for cat in categories]
-                if category_id not in valid_cat_ids:
-                    category_id = ""  # Reset if not valid
-
-    # Update bucket if requested
-    if (
-        update_bucket
-        and bucket_data is not None
-        and bucket_file_path is not None
-        and bucket_index is not None
-    ):
-        try:
-            # Update the bucket data
-            if summary_text:
-                bucket_data["summary"] = summary_text
-            if category_id:
-                bucket_data["category_id"] = category_id
-
-            # Read the file
-            with open(bucket_file_path, "r", encoding="utf-8") as f:
-                all_buckets = json.load(f)
-
-            # Update the bucket at the specified index
-            if 0 <= bucket_index < len(all_buckets):
-                all_buckets[bucket_index] = bucket_data
-
-                # Write back to the file
-                with open(bucket_file_path, "w", encoding="utf-8") as f:
-                    json.dump(all_buckets, f, indent=2)
-                print(f"Successfully updated bucket at index {bucket_index}")
-            else:
-                print(
-                    f"Error: bucket index {bucket_index} out of range (0-{len(all_buckets)-1})"
-                )
-        except Exception as e:
-            print(f"Error updating bucket: {e}")
-
-    return summary_text, category_id, suggested_category
-
-
-def generate_summary_and_category(bucket_data):
-    """Generate a summary and category for the current bucket."""
-    bucket_titles = bucket_data.get("titles", [])
-    bucket_ocr_texts = bucket_data.get("ocr_text", [])
-
-    # Skip processing if no data
-    if not bucket_titles and not bucket_ocr_texts:
-        return None, None
-
-    # Load categories
-    categories = load_categories()
-
-    # Prepare prompt
-    prompt_text = "Please provide a concise summary of computer activity based on the following window titles and detected text fragments. Focus on the primary tasks or topics.\n\n"
-    if bucket_titles:
-        prompt_text += (
-            "Window Titles:\n" + "\n".join([f'- "{t}"' for t in bucket_titles]) + "\n\n"
-        )
-    if bucket_ocr_texts:
-        prompt_text += (
-            "Detected Text (OCR Snippets):\n"
-            + "\n".join([f'- "{o}"' for o in bucket_ocr_texts])
-            + "\n\n"
-        )
-
-    # Add category options if available
-    if categories:
-        prompt_text += "Based on the activity, please categorize it into ONE of the following categories:\n\n"
-        for cat in categories:
-            prompt_text += (
-                f"- {cat.get('name')} ({cat.get('id')}): {cat.get('description')}\n"
-            )
-        prompt_text += "\nFirst, provide a concise summary of the activity.\nThen, on a new line after 'CATEGORY:', provide ONLY the category ID that best matches the activity."
-    else:
-        prompt_text += "Output only the summary text."
-
-    # Call LLM API
-    llm_response = _call_llm_api(prompt_text, "bucket summarization")
-    if not llm_response:
-        return None, None
-
-    # Parse response
-    summary_text = llm_response
-    category_id = None
-
-    if categories:
-        # Split on "CATEGORY:" if present
-        parts = llm_response.split("CATEGORY:")
-        if len(parts) > 1:
-            summary_text = parts[0].strip()
-            category_text = parts[1].strip()
-
-            # Get first word as category ID
-            if category_text:
-                category_id = (
-                    category_text.split()[0] if category_text.split() else None
-                )
-
-                # Validate category ID
-                valid_ids = [cat.get("id", "") for cat in categories]
-                if category_id not in valid_ids:
-                    category_id = None  # Reset if invalid
-
-    return summary_text, category_id
-
-
-def generate_summary_from_raw_with_llm(
-    bucket_titles: List[str],
-    bucket_ocr_texts: List[str],
-    allow_suggestions: bool = True,
-) -> Tuple[Optional[str], Optional[str], Optional[str]]:
-    """
-    Generate a summary, category, and optionally suggest a new category if needed.
-
-    Args:
-        bucket_titles: List of window titles
-        bucket_ocr_texts: List of OCR text snippets
-        allow_suggestions: Whether to allow the LLM to suggest new categories
-
-    Returns:
-        Tuple of (summary_text, category_id, suggested_category)
-    """
-    titles_to_send = (
-        [str(t) for t in bucket_titles if t and str(t).strip()] if bucket_titles else []
-    )
-    ocr_to_send = (
-        [str(o) for o in bucket_ocr_texts if o and str(o).strip()]
-        if bucket_ocr_texts
-        else []
-    )
-
-    text_parts = list(set(titles_to_send + ocr_to_send))  # Unique raw data points
-    if not text_parts:
-        print("No raw titles or OCR text available in this bucket")
-        return "", "", ""  # Return empty strings for summary, category, and suggestion
-
-    # Get available categories
-    categories = load_categories()
-
-    prompt_text = "Please provide a concise summary of computer activity based on the following window titles and detected text fragments. Focus on the primary tasks or topics.\n\n"
-
-    if titles_to_send:
-        prompt_text += (
-            "Window Titles:\n"
-            + "\n".join([f'- "{t}"' for t in titles_to_send])
-            + "\n\n"
-        )
-    if ocr_to_send:
-        prompt_text += (
-            "Detected Text (OCR Snippets):\n"
-            + "\n".join([f'- "{o}"' for o in ocr_to_send])
-            + "\n\n"
-        )
-
-    # Add category classification if categories exist
-    if categories:
-        prompt_text += "Based on the activity, please categorize it into ONE of the following categories:\n\n"
-        for cat in categories:
-            prompt_text += (
-                f"- {cat.get('name')} ({cat.get('id')}): {cat.get('description')}\n"
-            )
-
-        if allow_suggestions:
-            prompt_text += "\nIf none of these categories fit well, you can suggest a new category instead."
-            prompt_text += "\n\nFirst, provide a concise summary of the activity."
-            prompt_text += "\nThen, on a new line after 'CATEGORY:', provide ONLY the category ID that best matches the activity, or 'none' if no categories fit well."
-            prompt_text += "\nIf you answered 'none', then on a new line after 'SUGGESTION:', provide a suggested new category name and short description in the format 'name | description'."
-        else:
-            prompt_text += "\nFirst, provide a concise summary of the activity. Then, on a new line after 'CATEGORY:', provide ONLY the category ID that best matches the activity."
-    else:
-        prompt_text += "Output only the summary text."
-
-    llm_response = _call_llm_api(prompt_text, "raw data summarization")
-
-    if not llm_response:
-        return "", "", ""
-
-    # Parse response to extract summary, category, and suggestion
-    summary_text = llm_response
-    category_id = ""
-    suggested_category = ""
-
-    if categories:
-        parts = llm_response.split("CATEGORY:")
-        if len(parts) > 1:
-            summary_text = parts[0].strip()
-            category_text = parts[1].strip()
-
-            # Check for suggestion
-            suggestion_parts = category_text.split("SUGGESTION:")
-            if len(suggestion_parts) > 1:
-                category_text = suggestion_parts[0].strip()
-                suggested_category = suggestion_parts[1].strip()
-
-            # Extract just the category ID (assuming it's the first word after "CATEGORY:")
-            category_id = category_text.split()[0] if category_text.split() else ""
-
-            # Handle case where LLM says 'none'
-            if category_id.lower() == "none":
-                category_id = ""
-            else:
-                # Verify category ID exists
-                valid_cat_ids = [cat.get("id", "") for cat in categories]
-                if category_id not in valid_cat_ids:
-                    category_id = ""  # Reset if not valid
-
-    return summary_text, category_id, suggested_category
-
-
-def refine_summary_with_llm(
-    original_summary: str,
-    user_feedback: str,
-    current_category_id: str = "",
-    allow_suggestions: bool = True,
-) -> Tuple[Optional[str], Optional[str], Optional[str]]:
-    """
-    Refine an existing summary based on user feedback, preserving category if possible.
-
-    Args:
-        original_summary: The current summary text
-        user_feedback: User's feedback or input
-        current_category_id: Current category ID (if any)
-        allow_suggestions: Whether to allow the LLM to suggest new categories
-
-    Returns:
-        Tuple of (refined_summary, category_id, suggested_category)
-    """
-    categories = load_categories()
-    current_category_name = "Uncategorized"
-
-    if current_category_id:
-        cat_match = next(
-            (
-                cat.get("name", "Unknown")
-                for cat in categories
-                if cat.get("id") == current_category_id
-            ),
-            "Uncategorized",
-        )
-        current_category_name = cat_match
-
-    prompt = f"""Current summary of activity:
-"{original_summary}"
-
-User's feedback or additional details:
-"{user_feedback}"
-
-Based on the current summary and the user's input, please provide an improved and concise summary.
-If the user's input suggests a complete rewrite, then generate that new summary.
-Focus on integrating the user's points accurately.
-"""
-
-    if categories:
-        prompt += f"\nThe current activity is categorized as: {current_category_name} ({current_category_id})\n\n"
-        prompt += "Available categories:\n"
-        for cat in categories:
-            prompt += (
-                f"- {cat.get('name')} ({cat.get('id')}): {cat.get('description')}\n"
-            )
-
-        if allow_suggestions:
-            prompt += "\nIf none of these categories fit well, you can suggest a new category instead."
-            prompt += "\n\nFirst, provide the updated summary."
-            prompt += f"\nThen, on a new line after 'CATEGORY:', provide the category ID that best matches (either keep '{current_category_id}' or suggest a different one if appropriate), or 'none' if no categories fit well."
-            prompt += "\nIf you answered 'none', then on a new line after 'SUGGESTION:', provide a suggested new category name and short description in the format 'name | description'."
-        else:
-            prompt += "\nFirst, provide the updated summary. Then, on a new line after 'CATEGORY:', provide the category ID that best matches (either keep the current one or suggest a different one if appropriate)."
-    else:
-        prompt += "\nOutput only the new summary text."
-
-    llm_response = _call_llm_api(prompt, "refinement")
-
-    if not llm_response:
-        return None, current_category_id, ""  # Maintain current category if LLM fails
-
-    # Parse response to extract summary, category, and suggestion
-    refined_summary = llm_response
-    category_id = current_category_id
-    suggested_category = ""
-
-    if categories:
-        parts = llm_response.split("CATEGORY:")
-        if len(parts) > 1:
-            refined_summary = parts[0].strip()
-            category_text = parts[1].strip()
-
-            # Check for suggestion
-            suggestion_parts = category_text.split("SUGGESTION:")
-            if len(suggestion_parts) > 1:
-                category_text = suggestion_parts[0].strip()
-                suggested_category = suggestion_parts[1].strip()
-
-            # Extract category ID
-            category_id = (
-                category_text.split()[0]
-                if category_text.split()
-                else current_category_id
-            )
-
-            # Handle case where LLM says 'none'
-            if category_id.lower() == "none":
-                category_id = ""
-            else:
-                # Verify category ID exists
-                valid_cat_ids = [cat.get("id", "") for cat in categories]
-                if (
-                    category_id not in valid_cat_ids
-                    and category_id != current_category_id
-                ):
-                    category_id = current_category_id  # Keep current if invalid
-
-    return refined_summary, category_id, suggested_category
-
-
-# --- Visualization Functions ---
-def create_pie_chart(app_data: List[Dict[str, Any]]) -> go.Figure:
-    fig = go.Figure()
-    if not app_data:
-        fig.update_layout(title="App Usage Breakdown - No Data Available")
-        return fig
-
-    # Group small slices into "Other"
-    data_to_plot = sorted(app_data, key=lambda x: x.get("timeSpent", 0), reverse=True)
-    if len(data_to_plot) > 10:
-        top_apps = data_to_plot[:9]  # Show top 9
-        other_time = sum(app.get("timeSpent", 0) for app in data_to_plot[9:])
-        if other_time > 0:
-            top_apps.append({"appName": "Other Apps", "timeSpent": other_time})
-        data_to_plot = top_apps
-
-    labels = [
-        f"{app.get('appName','N/A')} ({app.get('timeSpent',0)//60}m)"
-        for app in data_to_plot
-    ]
-    values = [app.get("timeSpent", 0) for app in data_to_plot]
-
-    if not values or sum(values) == 0:  # Handle case where all values are zero
-        fig.update_layout(title="App Usage Breakdown - No Time Spent Data")
-        return fig
-
-    fig.add_trace(
-        go.Pie(
-            labels=labels,
-            values=values,
-            hole=0.3,
-            textinfo="percent+label",
-            hoverinfo="label+value",
-            textfont_size=12,
-            pull=[0.05] * len(labels),
-        )
-    )
-    fig.update_layout(
-        title_text="App Usage Breakdown", height=500, legend_title_text="Applications"
-    )
-    return fig
-
-
-def create_browser_chart(app_data: List[Dict[str, Any]]) -> Optional[go.Figure]:
-    if not app_data:
-        return None
-    browser_entries = []
-    for app in app_data:
-        name_lower = app.get("appName", "").lower()
-        if any(
-            b_keyword in name_lower
-            for b_keyword in ["chrome", "msedge", "edge", "firefox"]
-        ):
-            browser_type = (
-                "MS Edge"
-                if "msedge" in name_lower or "edge" in name_lower
-                else ("Google Chrome" if "chrome" in name_lower else "Firefox")
-            )
-            browser_entries.append(
-                {**app, "browserType": browser_type}
-            )  # Add type for coloring
-
-    if not browser_entries:
-        return None
-
-    df_browsers = pd.DataFrame(browser_entries).sort_values(
-        by=["browserType", "timeSpent"], ascending=[True, False]
-    )
-    df_browsers["timeSpentMinutesText"] = (df_browsers["timeSpent"] // 60).astype(
-        str
-    ) + "m"
-
-    fig = px.bar(
-        df_browsers,
-        x="appName",
-        y="timeSpent",
-        text="timeSpentMinutesText",
-        labels={
-            "appName": "Browser Instance / Profile",
-            "timeSpent": "Time Spent (seconds)",
-        },
-        title="Browser Usage Details",
-        color="browserType",
-        color_discrete_map={
-            "MS Edge": "#0078D4",
-            "Google Chrome": "#DB4437",
-            "Firefox": "#FF7139",
-        },
-    )
-    fig.update_layout(
-        xaxis_tickangle=-45,
-        height=450,
-        legend_title_text="Browser Type",
-        uniformtext_minsize=8,
-        uniformtext_mode="hide",
-    )
-    fig.update_traces(textposition="outside")
-    return fig
-
-
-def create_category_chart(
-    buckets: List[Dict[str, Any]], categories: List[Dict[str, str]]
-) -> Optional[go.Figure]:
-    """Create a pie chart showing time distribution by category."""
-    if not buckets or not categories:
-        return None
-
-    # Create a mapping of category ID to name
-    cat_id_to_name = {
-        cat.get("id", ""): cat.get("name", "Unknown") for cat in categories
-    }
-
-    # Count time per category
-    category_times = {"Uncategorized": 0}
-    for bucket in buckets:
-        cat_id = bucket.get("category_id", "")
-        cat_name = (
-            cat_id_to_name.get(cat_id, "Uncategorized") if cat_id else "Uncategorized"
-        )
-
-        # Calculate duration
-        start_time = pd.to_datetime(bucket.get("start", ""))
-        end_time = pd.to_datetime(bucket.get("end", bucket.get("start", "")))
-        if start_time and end_time:
-            duration = (end_time - start_time).total_seconds()
-            category_times[cat_name] = category_times.get(cat_name, 0) + duration
-
-    # Remove categories with zero time
-    category_times = {k: v for k, v in category_times.items() if v > 0}
-
-    if not category_times:
-        return None
-
-    # Create data for the chart
-    labels = [f"{cat} ({int(time//60)}m)" for cat, time in category_times.items()]
-    values = list(category_times.values())
-
-    fig = go.Figure()
-    fig.add_trace(
-        go.Pie(
-            labels=labels,
-            values=values,
-            hole=0.4,
-            textinfo="percent+label",
-            textfont_size=12,
-        )
-    )
-    fig.update_layout(title_text="Time Distribution by Category", height=450)
-    return fig
-
-
-# --- Activity Categories Manager Page ---
 def display_category_manager():
     st.title("🏆 Activity Categories Manager")
